project_postagram/postgram/requirements.txt
django==5.0.0
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-filter==21.1